Remove class CustomObjectKey. #176
