+++ models/src/main/java/io/sphere/sdk/customobjects/CustomObjectImpl.java
@@ -1,11 +1,15 @@
 package io.sphere.sdk.customobjects;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import io.sphere.sdk.models.DefaultModelViewImpl;
 
 import java.time.Instant;
 
+import static java.lang.String.format;
+
 class CustomObjectImpl<T> extends DefaultModelViewImpl<CustomObject<T>> implements CustomObject<T> {
+    private static final String keyContainerRegex = "[-_~.a-zA-Z0-9]+";
     private final String container;
     private final String key;
     private final T value;
@@ -30,5 +34,14 @@ class CustomObjectImpl<T> extends DefaultModelViewImpl<CustomObject<T>> implemen
     @Override
     public T getValue() {
         return value;
+    }
+
+    @JsonIgnore
+    static String validatedKey(final String name, final String keyOrContainer) {
+        if (!keyOrContainer.matches(keyContainerRegex)) {
+            final String message = format("The %s \"%s\" does not have the correct format. Key and container need to match %s.", name, keyOrContainer, keyContainerRegex);
+            throw new IllegalArgumentException(message);
+        }
+        return keyOrContainer;
     }
 }
